refactor(issue-133): move card subsystem files into card/ (no rename yet)
Creates web/src/card/ and relocates:
- physics/PhysicsCard.{tsx,css,test.tsx} → card/Card.{tsx,css,test.tsx}
- physics/PhysicsPage.{tsx,test.tsx}     → card/Page.{tsx,test.tsx}
- transitions/CardRegistry.{tsx,test.tsx} + CardRegistryStore.test.ts → card/
- transitions/PhysicsLayer.{tsx,test.tsx} → card/CardLayer.{tsx,test.tsx}
- transitions/PhysicsCardImpl.{tsx,test.tsx} → card/CardImpl.{tsx,test.tsx}
- transitions/NavCardContent.{tsx,test.tsx} → routes/blog/ (single caller)

This commit is path-only: filenames change but the React component
identifiers stay PhysicsX. The identifier rename wave follows in the
next commit so each step leaves typecheck and tests green.

All 575 tests pass; typecheck clean.

Refs #133.
