+++ web/src/layouts/CanvasLayout.tsx
@@ -7,7 +7,7 @@ import { FrameBar } from '../canvas/FrameBar'
 import { StringLayer } from '../canvas/StringLayer'
 import { useFrameEdge } from '../canvas/useFrameEdge'
 import { TransitionDirector } from '../transitions/TransitionDirector'
-import { PhysicsLayer } from '../transitions/PhysicsLayer'
+import { PhysicsLayer } from '../card/CardLayer'
 import { PageDefRegistryProvider } from '../transitions/PageDefRegistry'
 import { pageDefs } from '../routes/pageDefs'
 import { edges } from '../transitions/edges'
