+++ web/src/routes/Home.test.tsx
@@ -2,8 +2,8 @@ import { describe, test, expect, vi } from 'vitest'
 import { render } from '@testing-library/react'
 import Home, { pageDef } from './Home'
 import { PhysicsProvider } from '../physics/PhysicsContext'
-import { CardRegistryProvider } from '../transitions/CardRegistry'
-import { PhysicsLayer } from '../transitions/PhysicsLayer'
+import { CardRegistryProvider } from '../card/CardRegistry'
+import { PhysicsLayer } from '../card/CardLayer'
 
 vi.mock('../canvas/flip', () => ({
     flipMorph: vi.fn(),
