+++ web/src/routes/blog/BlogIndex.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useMemo, useRef } from 'react'
 import { Link } from 'react-router-dom'
-import { PhysicsPage, type CardContent } from '../../physics/PhysicsPage'
+import { PhysicsPage, type CardContent } from '../../card/Page'
 import { measure } from '../../text/measure'
 import { getPosts } from '../../blog/posts'
 import {
@@ -19,7 +19,7 @@ import {
     NAV_CARD_H,
     NAV_TOP_INSET,
 } from '../../layout/sectionLayout'
-import { NavCardContent } from '../../transitions/NavCardContent'
+import { NavCardContent } from './NavCardContent'
 import type { PageDef } from '../PageDef'
 import type { CardSpec } from '../../physics/PageSpec'
 import type { Post } from '../../blog/types'
