+++ web/src/routes/blog/BlogPost.tsx
@@ -1,6 +1,6 @@
 import { useMemo, useState, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
-import { PhysicsCard } from '../../physics/PhysicsCard'
+import { PhysicsCard } from '../../card/Card'
 import { useRegisterPageDef } from '../../transitions/PageDefRegistry'
 import { getPosts } from '../../blog/posts'
 import { mdxComponents } from '../../blog/components/mdx-components'
