+++ web/src/routes/blog/NavCardContent.tsx
